72 Added UpgradeRepository and UpgradeRecipeRepository

+++ server/src/main/java/de/pnp/manager/component/inventory/equipment/interfaces/IDamageableEquipment.java
@@ -1,15 +1,22 @@
 package de.pnp.manager.component.inventory.equipment.interfaces;
 
+/**
+ * Represents {@link IEquipment} that can be damaged.
+ */
 public interface IDamageableEquipment extends IEquipment {
 
-    void applyWear(int wear);
+    /**
+     * Applies wear to the equipment.
+     */
+    void applyWear(float wear);
 
+    /**
+     * Completely repairs the equipment.
+     */
     void repair();
 
     /**
      * Returns a float [0,1] representing the current durability in percent.
-     *
-     * @return
      */
     float getRelativeDurability();
 }
